+++ jvm/src/main/scala/poolmate/Store.scala
@@ -220,7 +220,7 @@ final class Store(conf: Config,
   def listSupplies(): List[Supply] =
     DB readOnly { implicit session =>
       sql"select * from supply order by purchased desc"
-        .map(rs => Supply(rs.long("id"), rs.long("pool_id"), rs.int("purchased"), rs.string("item"), rs.double("amount"), rs.string("unit"), rs.int("cost")))
+        .map(rs => Supply(rs.long("id"), rs.long("pool_id"), rs.int("purchased"), rs.string("item"), rs.int("amount"), rs.string("unit"), rs.int("cost")))
         .list()
     }
